move scan results in other files
